Finish drafting APIs, next is testing and debugging

+++ routes/review.js
@@ -4,29 +4,21 @@ const mongoose = require('mongoose')
 const User = mongoose.model('User')
 const Food = mongoose.model('Food')
 const Review = mongoose.model('Review')
+import { createNotification } from './notification'
 
 // GET - Get Review
 router.get('/review/:foodname', auth.optional, async (req, res, next) => {
   try {
-    let authUser = await User.findOne({ sub: req.user.sub })
-    let review = await Review.findOne({ foodname: req.params.foodname })
-      .populate('account', 'username image')
-      .populate('tags', 'name')
-      .populate('comments')
-    return res.json({
-      isFollowing: authUser.isFollowing(review.account),
-      review: {
-        ...review,
-        isLiked: authUser.isLiked(review._id),
-        isSaved: authUser.isSaved(review.food._id)
-      },
-      comments: review.comments.map(comment => {
-        return {
-          ...comment,
-          isLiked: authUser.isLikedComment(comment._id)
-        }
-      })
-    })
+    let [user, review] = await Promise.all([
+      User.findOne({ sub: req.user.sub }),
+      Review.findOne({ foodname: req.params.foodname })
+        .populate('account', 'username image')
+        .populate('tags', 'name')
+        .populate('comments')
+    ])
+    if(!user || !review) return res.sendStatus(401)
+    return res.json({ review: review.getReview(user) })
+
   } catch (err) {
     console.log(err)
     next(err)
@@ -63,7 +55,15 @@ router.post('/review', auth.required, async (req, res, next) => {
     // save both review and food
     review.food = food._id
     await Promise.all([review.save(), food.save()])
-    return res.json({ review: review })
+
+    res.json({ review: review.getReview(user) })
+
+    // Create notifications
+    user.followers.forEach(async followerId => {
+      await createNotification('review', review._id, user._id, followerId)
+    })
+
+    return
 
   } catch (err) {
     console.log(err)
@@ -76,10 +76,11 @@ router.post('/review', auth.required, async (req, res, next) => {
 router.post('/review/:foodname', auth.required, async (req, res, next) => {
   try {
     let user = await User.findOne({ sub: req.user.sub })
-    if (!user || !req.body.review) return res.sendStatus(401)
+    if (!user || !req.body.review
+      || req.body.review.foodname !== req.params.foodname) return res.sendStatus(401)
 
     let r = req.body.review
-    let review = await Review.findOne({ foodname: req.params.foodname })
+    let review = await Review.findOne({ account: user._id, foodname: req.params.foodname })
       .select('-comments -likesCount -flaggedCount')
     if (!review) return res.sendStatus(404)
     if (!r.foodTitle || !r.address || !r.photos || !r.tags
@@ -91,7 +92,9 @@ router.post('/review/:foodname', auth.required, async (req, res, next) => {
 
     // Update Food
     let food = await Food.findById(review.food)
-    if (food.foodTitle !== r.foodTitle || food.address !== r.address) {
+    if (food.foodTitle === r.foodTitle && food.address === r.address) {
+      await food.setDetails(review, oldReview)
+    } else {
       // remove review from old food
       food.removeReview(review)
       if (food.reviews.length <= 0) await Food.findByIdAndDelete(food._id)
@@ -101,8 +104,6 @@ router.post('/review/:foodname', auth.required, async (req, res, next) => {
       food = await Food.findOne({ foodTitle: r.foodTitle, address: r.address })
       // if found, add review to found food
       if (food) food.setDetails(review, null)
-    } else {
-      await food.setDetails(review, oldReview)
     }
 
     if (!food) {
@@ -113,7 +114,7 @@ router.post('/review/:foodname', auth.required, async (req, res, next) => {
 
     // save both review and food
     await Promise.all([review.save(), food.save()])
-    return res.json({ review: review })
+    return res.json({ review: review.getReview(user) })
 
   } catch (err) {
     console.log(err)
@@ -126,11 +127,10 @@ router.post('/review/:foodname', auth.required, async (req, res, next) => {
 router.delete('/review/:foodname', auth.required, async (req, res, next) => {
   try {
     let user = await User.findOne({ sub: req.user.sub })
-    if (!user || !req.body.review) return res.sendStatus(401)
+    if (!user || !req.body.review
+      || req.body.review.foodname !== req.params.foodname) return res.sendStatus(401)
 
-    let r = req.body.review
-    let review = await Review.findOne({ foodname: req.params.foodname })
-      .select('-comments -likesCount -flaggedCount')
+    let review = await Review.findOne({ account: user._id, foodname: req.params.foodname })
     if (!review) return res.sendStatus(404)
 
     await Promise.all([review.setTags([]), review.deleteComments()])
@@ -143,7 +143,7 @@ router.delete('/review/:foodname', auth.required, async (req, res, next) => {
       else await food.save()
     }
 
-    await review.save()
+    await Review.findByIdAndDelete(review._id)
     return res.json({ review: {} })
 
   } catch (err) {
@@ -152,5 +152,70 @@ router.delete('/review/:foodname', auth.required, async (req, res, next) => {
   }
 })
 
+
+// GET - All Current Reviews
+router.get('/reviews', auth.optional, async (req, res, next) => {
+  try {
+    let query = {}, options = {}
+    if (req.query.keywords) {
+      query = { $text: { $search: req.query.keywords } }
+      options = { score: { $meta: "textScore" } }
+    }
+
+    let limit = 12
+    let offset = 0
+    if (typeof req.query.offset !== 'undefined') offset = req.query.offset
+    let user = req.user ? await User.findOne({ sub: req.user.sub }) : null
+    let reviews = await Review.find(query, options)
+      .limit(Number(limit))
+      .skip(Number(offset))
+      .sort(options)
+
+    return res.json({
+      reviews: reviews.map(function (review) {
+        return review.getReviewBasic(user)
+      })
+    })
+
+  } catch (err) {
+    console.log(err)
+    next(err)
+  }
+})
+
+
+// Like
+router.put('/review/like/:reviewId', auth.required, async (req, res, next) => {
+  try {
+    let [user, review] = await Promise.all([
+      User.findOne({ sub: req.user.sub }),
+      Review.findById(req.params.reviewId)
+    ])
+    if (!user || !review) return res.sendStatus(401)
+    await user.like(review)
+    return res.json({ isLiked: user.isLiked(review._id), likesCount: review.likesCount })
+
+  } catch (err) {
+    console.log(err)
+    next(err)
+  }
+})
+
+// Unlike
+router.put('/review/unlike/:reviewId', auth.required, async (req, res, next) => {
+  try {
+    let [user, review] = await Promise.all([
+      User.findOne({ sub: req.user.sub }),
+      Review.findById(req.params.reviewId)
+    ])
+    if (!user || !review) return res.sendStatus(401)
+    await user.unlike(review)
+    return res.json({ isLiked: user.isLiked(review._id), likesCount: review.likesCount })
+
+  } catch (err) {
+    console.log(err)
+    next(err)
+  }
+})
 
 module.exports = router
